requirements_saas.txt
# Django Core (Latest LTS)
Django==5.1.3
django-environ==0.11.2
django-cors-headers==4.4.0

# Django REST Framework
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular[sidecar]==0.27.2

# Authentication & Security
django-allauth==0.63.6
dj-rest-auth==6.0.0
django-oauth-toolkit==3.0.1
PyJWT==2.9.0
django-two-factor-auth==1.16.0
qrcode==7.4.2
django-otp==1.5.2

# Database
psycopg2-binary==2.9.10
django-redis==5.4.0

# Billing & Payments
stripe==8.7.0
razorpay==1.4.2

# API Rate Limiting & Throttling
django-ratelimit==4.1.0
djangorestframework-api-key==3.0.0

# Monitoring & Logging
django-extensions==3.2.3
celery==5.3.6
redis==5.0.3
sentry-sdk==1.45.0

# Security
django-csp==3.8

# Utils
requests==2.32.3
Pillow==10.4.0
django-storages==1.14.4
boto3==1.35.36

# Development
django-debug-toolbar==4.4.6
factory-boy==3.3.1

# Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Original GST API dependencies
Flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
flask-restx==1.3.0
typing-extensions>=4.9.0
uuid==1.30